requirements.txt
streamlit==1.38.0
torch==2.4.1
torchvision==0.19.1
opencv-python-headless==4.10.0.84
pillow==10.4.0
numpy==1.26.4
